refactor : minor changes in all tasks
